requirements.txt
SQLAlchemy>=1.4.37
graphene_sqlalchemy>=2.3.0
Flask-GraphQL>=2.0.1
Flask>=2.1.2
flask_sqlalchemy>=2.5.1
mysqlclient>=2.1.0
